Move 'property' to private folder
